portal/backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
msal==1.26.0

# Database
tinydb==4.8.0

# Redis / Queue
redis==5.0.1
rq==1.16.0

# Azure
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# WebSocket
websockets==12.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
Jinja2==3.1.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
